+++ lib/prefab/options.rb
@@ -32,14 +32,26 @@ module Prefab
       progname = (progname.nil? || progname.empty?) ? path : "#{progname}: #{path}"
 
       formatted_log_context = log_context.sort.map do |k, v|
-        v.nil? || ((v.is_a? Array) && v.empty?) ? nil : "#{k}=#{v}"
-      end.compact.join(" ")
+        "#{k}=#{v}"
+      end.join(" ")
       "#{severity.ljust(5)} #{datetime}:#{' ' if progname}#{progname} #{msg}#{log_context.any? ? " " + formatted_log_context : ""}\n"
     }
 
     JSON_LOG_FORMATTER = proc { |data|
       log_context = data.delete(:log_context)
       data.merge(log_context).compact.to_json << "\n"
+    }
+
+    COMPACT_LOG_FORMATTER = proc { |data|
+      severity = data[:severity]
+      msg = data[:message]
+      log_context = data[:log_context]
+      log_context["path"] = data[:path] || ""
+
+      formatted_log_context = log_context.sort.map do |k, v|
+        "#{k}=#{v}"
+      end.join(" ")
+      "#{severity.ljust(5)} #{msg&.strip} #{formatted_log_context}\n"
     }
 
     module ON_INITIALIZATION_FAILURE